Settings: Move "Adaptive Playback" option to Sound Settings.
 * Our gestures category looks bloated with features.
   Move this to Sound Settings where it feels better and more homely.
