+++ res/xml/gestures.xml
@@ -99,10 +99,4 @@
         android:fragment="com.android.settings.gestures.PreventRingingGestureSettings"
         settings:controller="com.android.settings.gestures.PreventRingingParentPreferenceController" />
 
-    <Preference
-        android:key="gesture_adaptive_playback_summary"
-        android:title="@string/adaptive_playback_screen_title"
-        android:fragment="com.android.settings.gestures.AdaptivePlaybackGestureSettings"
-        settings:controller="com.android.settings.gestures.AdaptivePlaybackParentPreferenceController" />
-
 </PreferenceScreen>
